fixing tree
