+++ wig/bigWig/io_test.go
